refactor: 사용자 인증 객체 패키지 분리
